Corrección de variables

+++ tema2SamuelJimenez/src/main/java/algoritmossecuenciales/Ej05.java
@@ -21,7 +21,7 @@ public class Ej05 {
         Scanner sc = new Scanner(System.in);
 
         final int HORASCLASE = 30;
-        int horasImpartidas, grupos, profesores;
+        double horasImpartidas, grupos, profesores;
 
         //HORASCLASE*GRUPOS/HORASIMPARTIDAS = PROFESORES.
         System.out.println("Introduzca las horas que impartirá cada profesor");
@@ -35,11 +35,11 @@ public class Ej05 {
 
         //Hacemos la opción de JOptionPane
         String horasTxt = JOptionPane.showInputDialog("¿Cuantas horas impartirá cada profesor?");
-        int horasProfesor = Integer.parseInt(horasTxt);
+        double horasProfesor = Double.parseDouble(horasTxt);
         String gruposTxt = JOptionPane.showInputDialog("¿Cuántos grupos tiene el centro?");
-        int gruposClase = Integer.parseInt(gruposTxt);
+        double gruposClase = Double.parseDouble(gruposTxt);
 
-        int profesoresTotales = (HORASCLASE * gruposClase) / horasProfesor;
+        double profesoresTotales = (HORASCLASE * gruposClase) / horasProfesor;
 
         JOptionPane.showMessageDialog(null, "Para este centro harán falta " + (int) Math.ceil(profesoresTotales) + " profesores.");
